b2g/components/test/mochitest/test_screenshot.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1136784
-->
<head>
  <meta charset="utf-8">
  <title>Screenshot Test</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1136784">Screenshot.jsm</a>
<script type="application/javascript">

"use strict";

var gUrl = SimpleTest.getTestFileURL("screenshot_helper.js");
var gScript = SpecialPowers.loadChromeScript(gUrl);

SimpleTest.waitForExplicitFinish();
gScript.addMessageListener("finish", function () {
  SimpleTest.ok(true, "chrome test script finished");
  gScript.destroy();
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
